Renamed some files and finished api CRUD methods
