+++ starter-code/models/Place.model.js
@@ -5,7 +5,7 @@ module.exports = mongoose.model('Place', new Schema({
   name: String,
   type: {
     type: String,
-    enum: ['Coffee shop', 'bookstore'],
+    enum: ['Coffee shop', 'Bookstore'],
   },
 },
 {
